// src/integrations/cms/generated/cms-schema.d.ts
export interface components {
    schemas: {
// ...
        ApiPromotionPromotionDocument: {
// ...
            content: unknown[];
// ...
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
// ...
    "article/get/articles_by_id": {
// ...
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        data: {
                            /**
                             * Format: uuid
                             * @description The document ID, represented by a UUID
                             */
                            documentId: string;
                            id: string | number;
                            /** @description A string field */
                            title?: string;
                            /** @description A text field */
                            description?: string;
                            /** @description A blocks field */
                            content: unknown[];
                            /** @description A UID field */
                            slug?: string;
                            /** @description A datetime field */
                            publish_date?: string;
                            /**
                             * @description An enum field
                             * @enum {string}
                             */
                            articleType?: "news" | "blog" | "press_release" | "announcement";
                            /**
                             * @description An enum field
                             * @default all
                             * @enum {string}
                             */
                            customerSegment: "individual" | "corporate" | "priority" | "all";
                            /** @description A boolean field */
                            isFeatured?: boolean | null;
                            /** @description A datetime field */
                            createdAt?: string;
                            /** @description A datetime field */
                            updatedAt?: string;
                            /**
                             * @description A datetime field
                             * @default 2026-06-04T06:49:03.492Z
                             */
                            publishedAt: string;
                            /** @description A string field */
                            locale?: string;
                            /** @description A media field */
                            cover?: components["schemas"]["PluginUploadFileDocument"];
                            /** @description A media field */
                            thumbnail?: components["schemas"]["PluginUploadFileDocument"];
                            /** @description A media field */
                            attachments?: components["schemas"]["PluginUploadFileDocument"][];
                            /** @description A relational field */
                            tags?: components["schemas"]["ApiTagTagDocument"][];
                            /** @description A dynamic zone field */
                            blocks?: unknown[];
                            /** @description A component field */
                            seo?: components["schemas"]["SharedSeoEntry"];
                            /** @description A relational field */
                            author?: components["schemas"]["ApiAuthorAuthorDocument"];
                            /** @description A relational field */
                            category?: components["schemas"]["ApiCategoryCategoryDocument"];
                            /** @description A relational field */
                            relatedProducts?: components["schemas"]["ApiProductProductDocument"][];
                            /** @description A relational field */
                            relatedArticles?: components["schemas"]["ApiArticleArticleDocument"][];
                            /** @description A relational field */
                            readonly localizations?: components["schemas"]["ApiArticleArticleDocument"][];
                        };
                    };
                };
            };
// ...
        };
    };
// ...
}